credfloat-assignment/connector/requirements.txt
pyodbc>=5.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pystray>=0.19.5
Pillow>=10.0.0
pyinstaller>=6.3.0
